[inquiry] update

+++ src/main/java/hello/fclover/domain/Goods.java
@@ -1,5 +1,6 @@
 package hello.fclover.domain;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -7,7 +8,7 @@ import lombok.Data;
 
 @Data
 public class Goods {
-    private Long goodsNo;              // 상품 일련번호
+    private BigInteger goodsNo;              // 상품 일련번호
     private Long sellerNo;             // 판매자 번호
     private Integer cateNo;            // 카테고리 번호
     private String goodsName;          // 상품 명
@@ -21,5 +22,7 @@ public class Goods {
     private LocalDateTime goodsDate;   // 상품 등록일
     private int goodsPageCount;        // 상품 페이지 수
     private String goodsBookSize;      // 상품 크기
+
+
 }
